Update the UI to respond to the dynamically calculated zoom limits
These changes feed the dynamically calculated zoom limits into the Map
component.

The addition of the "key" attribute is to get around the fact that
minZoom and maxZoom are not "dynamically" updated by the React-Leaflet
bindings. This forces react to completely recreate the map component. It
is not a very efficient choice, but it will do for now.